+++ src/Algorithms/Algos/affinemodel.h
@@ -8,7 +8,7 @@ namespace astro
 {
 class ASTRO_ALGORITHMS_EXPORT AffineModel
 {
-    Eigen::Matrix<float, Eigen::Dynamic, Eigen::Dynamic> m_A;
+    Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic> m_A;
 
 public:
     AffineModel();
@@ -16,12 +16,12 @@ public:
     ~AffineModel();
     AffineModel& operator=(const AffineModel&);
 
-    Eigen::Matrix<float, Eigen::Dynamic, Eigen::Dynamic> getData() const;
+    Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic> getData() const;
 
-    void fit(const Eigen::Matrix<float, Eigen::Dynamic, Eigen::Dynamic>& X,
-             const Eigen::Matrix<float, Eigen::Dynamic, Eigen::Dynamic>& Y);
+    void fit(const Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic>& X,
+             const Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic>& Y);
 
-    Eigen::Matrix<float, Eigen::Dynamic, Eigen::Dynamic>
-    predict(const Eigen::Matrix<float, Eigen::Dynamic, Eigen::Dynamic>& X) const;
+    Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic>
+    predict(const Eigen::Matrix<double, Eigen::Dynamic, Eigen::Dynamic>& X) const;
 };
 } // namespace astro
